Release v5.0.0a13 (imported from PyPI)

+++ otree/templating/filters.py
@@ -71,7 +71,7 @@ def json(val):
 
 @register
 def to0(val):
-    return round(val, 0)
+    return int(round(val, 0))
 
 
 @register
